requirements.txt
# PIP install file
pyyaml
pillow
matplotlib
discord.py
pyspellchecker
pytz
requests
peewee
openfoodfacts